Fixes #39: http-collector non-blocking collect
`http.collector.Collect` was blocking, thus interfering with application
performance when zipkin is slow, not responding or down. The collect
function could take up to `c.client.Timeout`.

This fix makes the Collect function asynchronous, sending all the
spans in the batch, and remove them from the batch once succeeds.

Add some tests to verify the batching and timouting issues.