Closes #41: collector-http error logs are too verbose
Add log error interval to collector-http, which reduces the amount
of logs reported by the collector in case that remote Zipkin is down.
The logging of the error will happen when the error is changed
or when the configured time elapsed.
A restore connection log will appear when the connection
is restored, and the logging interval was defined.