Add an Observer API and its implementation
Metrics are useful to gain insights in a distributed application. But
there can be a lot of metrics in different domains. Adding such metrics
from one domain (metrics exporter pkg) into zipkin is not good for
cross platform compatibility.
This commit adds a new observer and span observer API which defines a
standard interface to add such metrics. To expose the metrics, one would
have to implement the observer interface in the metrics
exporter. zipkin will need to implement this interface and install
callbacks for methods such as StartSpan, SetOperationName, SetTag and
Finish. The registered callbacks would then be called on the span events
if an observer is created.

This is based on the work done by Yuri here :
https://github.com/uber/jaeger-client-go/pull/94

Signed-off-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>
Signed-off-by: Hemant Kumar <hemant@linux.vnet.ibm.com>