fixed: use non blocking channel in HTTPCollector.Collect()
This patch makes the span collection channel buffered, and the Collect() method
to use a non blocking send. The buffer size of the channel is equal
to the backlog size. It also removes the actual async HTTP post.

Previously, when the application was under heavy load, the number of send() go routines
was skyrocketing, reaching the -race goroutine limit of 8192 in less than a 10sec.

I see that the scribe collector is using the same system, I would be happy to convert it too
if this patch gets accepted.